- dont show 0 in port value by default - wired up edit button and started loading saved data into fragments - filter out invalid characters in hostname field

+++ PortKnocker/src/com/xargsgrep/portknocker/manager/HostDataManager.java
@@ -50,7 +50,7 @@ public class HostDataManager {
 		return hosts;
 	}
 	
-	public Host getHost(int hostId) {
+	public Host getHost(long hostId) {
 		SQLiteDatabase database = getReadableDatabase();
 		
 		String hostSelection = String.format("%s = ?", DatabaseManager.HOST_ID_COLUMN);
